feat(web): multi answer mode ui
